Move common integration test code to integration/package-operator/helpers.go
