+++ integration/package-operator/environment.go
@@ -1,5 +1,3 @@
-//go:build integration
-
 package packageoperator
 
 import (
